+++ pom.xml
@@ -149,7 +149,6 @@
         <module>twitter-eventlistener-plugin</module>
         <module>presto-twitter-server</module>
         <module>presto-twitter-functions</module>
-        <module>presto-twitter-gateway</module>
         <module>presto-twitter-server-rpm</module>
     </modules>
