+++ presto-hive-metastore/src/main/java/com/facebook/presto/twitter/hive/MetastoreStaticClusterModule.java
@@ -14,19 +14,21 @@
 package com.facebook.presto.twitter.hive;
 
 import com.facebook.presto.hive.metastore.thrift.HiveCluster;
+import com.facebook.presto.hive.metastore.thrift.StaticHiveCluster;
+import com.facebook.presto.hive.metastore.thrift.StaticMetastoreConfig;
 import com.google.inject.Binder;
 import com.google.inject.Module;
 import com.google.inject.Scopes;
 
-import static io.airlift.configuration.ConfigBinder.configBinder;
+import static com.facebook.airlift.configuration.ConfigBinder.configBinder;
 
-public class MetastoreZkDiscoveryBasedModule
+public class MetastoreStaticClusterModule
         implements Module
 {
     @Override
     public void configure(Binder binder)
     {
-        binder.bind(HiveCluster.class).to(ZookeeperServersetHiveCluster.class).in(Scopes.SINGLETON);
-        configBinder(binder).bindConfig(ZookeeperServersetMetastoreConfig.class);
+        binder.bind(HiveCluster.class).to(StaticHiveCluster.class).in(Scopes.SINGLETON);
+        configBinder(binder).bindConfig(StaticMetastoreConfig.class);
     }
 }
