+++ presto-hive-metastore/src/main/java/com/facebook/presto/twitter/hive/thrift/ThriftGeneralInputFormat.java
@@ -29,10 +29,10 @@ import org.apache.hadoop.mapred.Reporter;
 
 import java.io.IOException;
 
-import static com.facebook.presto.hive.HiveErrorCode.HIVE_INVALID_METADATA;
-import static com.facebook.presto.hive.HiveUtil.checkCondition;
-import static com.facebook.presto.hive.HiveUtil.getLzopIndexPath;
-import static com.facebook.presto.hive.HiveUtil.isLzopCompressedFile;
+import static com.facebook.presto.hive.MetastoreErrorCode.HIVE_INVALID_METADATA;
+import static com.facebook.presto.hive.metastore.MetastoreUtil.checkCondition;
+import static com.facebook.presto.twitter.hive.thrift.LzoThriftUtil.getLzopIndexPath;
+import static com.facebook.presto.twitter.hive.thrift.LzoThriftUtil.isLzopCompressedFile;
 import static java.lang.String.format;
 import static org.apache.hadoop.hive.serde.Constants.SERIALIZATION_CLASS;
