+++ presto-hive-metastore/src/test/java/com/facebook/presto/twitter/hive/util/TestUtils.java
