+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveUtil.java
@@ -50,7 +50,6 @@ import io.airlift.slice.Slices;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.fs.PathFilter;
 import org.apache.hadoop.hive.common.JavaUtils;
 import org.apache.hadoop.hive.ql.exec.Utilities;
 import org.apache.hadoop.hive.ql.io.SymlinkTextInputFormat;
@@ -142,7 +141,6 @@ import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Iterables.filter;
 import static com.google.common.collect.Lists.newArrayList;
 import static com.google.common.collect.Lists.transform;
-import static com.hadoop.compression.lzo.LzoIndex.LZO_INDEX_SUFFIX;
 import static java.lang.Byte.parseByte;
 import static java.lang.Double.parseDouble;
 import static java.lang.Float.floatToRawIntBits;
@@ -178,22 +176,6 @@ public final class HiveUtil
     private static final int DECIMAL_SCALE_GROUP = 2;
 
     private static final String BIG_DECIMAL_POSTFIX = "BD";
-
-    private static final PathFilter LZOP_DEFAULT_SUFFIX_FILTER = new PathFilter() {
-        @Override
-        public boolean accept(Path path)
-        {
-            return path.toString().endsWith(".lzo");
-        }
-    };
-
-    private static final PathFilter LZOP_INDEX_DEFAULT_SUFFIX_FILTER = new PathFilter() {
-        @Override
-        public boolean accept(Path path)
-        {
-            return path.toString().endsWith(".lzo.index");
-        }
-    };
 
     static {
         DateTimeParser[] timestampWithoutTimeZoneParser = {
@@ -385,21 +367,6 @@ public final class HiveUtil
         catch (InvocationTargetException | IllegalAccessException e) {
             throw new RuntimeException(e);
         }
-    }
-
-    public static boolean isLzopCompressedFile(Path filePath)
-    {
-        return LZOP_DEFAULT_SUFFIX_FILTER.accept(filePath);
-    }
-
-    public static boolean isLzopIndexFile(Path filePath)
-    {
-        return LZOP_INDEX_DEFAULT_SUFFIX_FILTER.accept(filePath);
-    }
-
-    public static Path getLzopIndexPath(Path lzoPath)
-    {
-        return lzoPath.suffix(LZO_INDEX_SUFFIX);
     }
 
     public static StructObjectInspector getTableObjectInspector(@SuppressWarnings("deprecation") Deserializer deserializer)
