+++ presto-hive/src/main/java/com/facebook/presto/twitter/hive/thrift/ThriftHiveRecordCursorProvider.java
@@ -37,13 +37,13 @@ import java.util.Optional;
 import java.util.Properties;
 import java.util.Set;
 
-import static com.facebook.presto.hive.HiveErrorCode.HIVE_INVALID_METADATA;
 import static com.facebook.presto.hive.HiveStorageFormat.THRIFTBINARY;
-import static com.facebook.presto.hive.HiveUtil.checkCondition;
 import static com.facebook.presto.hive.HiveUtil.createRecordReader;
 import static com.facebook.presto.hive.HiveUtil.getDeserializerClassName;
-import static com.facebook.presto.hive.HiveUtil.getLzopIndexPath;
-import static com.facebook.presto.hive.HiveUtil.isLzopCompressedFile;
+import static com.facebook.presto.hive.MetastoreErrorCode.HIVE_INVALID_METADATA;
+import static com.facebook.presto.hive.metastore.MetastoreUtil.checkCondition;
+import static com.facebook.presto.twitter.hive.thrift.LzoThriftUtil.getLzopIndexPath;
+import static com.facebook.presto.twitter.hive.thrift.LzoThriftUtil.isLzopCompressedFile;
 import static java.util.Objects.requireNonNull;
 
 public class ThriftHiveRecordCursorProvider
