--- presto-twitter-gateway/etc/config.properties
@@ -1,2 +1,0 @@
-node.environment=test
-http-server.http.port=8088
