--- presto-twitter-gateway/src/main/java/com/twitter/presto/gateway/PrestoGateway.java
@@ -1,67 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.twitter.presto.gateway;
-
-import com.google.common.collect.ImmutableList;
-import com.google.inject.Module;
-import io.airlift.bootstrap.Bootstrap;
-import io.airlift.event.client.EventModule;
-import io.airlift.http.server.HttpServerModule;
-import io.airlift.jaxrs.JaxrsModule;
-import io.airlift.jmx.JmxHttpModule;
-import io.airlift.jmx.JmxModule;
-import io.airlift.json.JsonModule;
-import io.airlift.log.LogJmxModule;
-import io.airlift.log.Logger;
-import io.airlift.node.NodeModule;
-import io.airlift.tracetoken.TraceTokenModule;
-import org.weakref.jmx.guice.MBeanModule;
-
-public class PrestoGateway
-{
-    private PrestoGateway() {}
-
-    public static void start(Module... extraModules)
-    {
-        Bootstrap app = new Bootstrap(ImmutableList.<Module>builder()
-                .add(new NodeModule())
-                .add(new HttpServerModule())
-                .add(new JsonModule())
-                .add(new JaxrsModule(true))
-                .add(new MBeanModule())
-                .add(new JmxModule())
-                .add(new JmxHttpModule())
-                .add(new TraceTokenModule())
-                .add(new LogJmxModule())
-                .add(new EventModule())
-                .add(new GatewayModule())
-                .add(extraModules)
-                .build());
-
-        Logger log = Logger.get(PrestoGateway.class);
-        try {
-            app.strictConfig().initialize();
-            log.info("======== SERVER STARTED ========");
-        }
-        catch (Throwable t) {
-            log.error(t);
-            System.exit(1);
-        }
-    }
-
-    public static void main(String[] args)
-    {
-        start();
-    }
-}
