--- presto-twitter-gateway/src/main/java/com/twitter/presto/gateway/cluster/ClusterManager.java
@@ -1,26 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.twitter.presto.gateway.cluster;
-
-import com.twitter.presto.gateway.query.QueryCategory;
-
-import java.net.URI;
-import java.util.List;
-
-public interface ClusterManager
-{
-    List<URI> getAllClusters();
-
-    List<URI> getClusters(QueryCategory category);
-}
