+++ presto-twitter-server-rpm/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.facebook.presto</groupId>
         <artifactId>presto-root</artifactId>
-        <version>0.228</version>
+        <version>0.230</version>
     </parent>
 
     <artifactId>presto-twitter-server-rpm</artifactId>
@@ -27,7 +27,7 @@
         <dependency>
             <groupId>com.facebook.presto</groupId>
             <artifactId>presto-twitter-server</artifactId>
-            <version>0.228</version>
+            <version>0.230</version>
             <type>tar.gz</type>
             <scope>runtime</scope>
         </dependency>
