+++ presto-twitter-server/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.facebook.presto</groupId>
         <artifactId>presto-root</artifactId>
-        <version>0.228</version>
+        <version>0.230</version>
     </parent>
 
     <artifactId>presto-twitter-server</artifactId>
