+++ twitter-eventlistener-plugin/src/main/java/com/twitter/presto/plugin/eventlistener/TwitterEventListenerConfig.java
@@ -13,7 +13,7 @@
  */
 package com.twitter.presto.plugin.eventlistener;
 
-import io.airlift.configuration.Config;
+import com.facebook.airlift.configuration.Config;
 
 public class TwitterEventListenerConfig
 {
