+++ twitter-eventlistener-plugin/src/test/java/com/twitter/presto/plugin/eventlistener/TestTwitterEventListenerPlugin.java
@@ -22,8 +22,8 @@ import org.testng.annotations.Test;
 import java.util.Map;
 import java.util.ServiceLoader;
 
+import static com.facebook.airlift.testing.Assertions.assertInstanceOf;
 import static com.google.common.collect.Iterables.getOnlyElement;
-import static io.airlift.testing.Assertions.assertInstanceOf;
 import static org.testng.Assert.assertNotNull;
 
 public class TestTwitterEventListenerPlugin
